requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
networkx
python-dotenv
